backend/voice-service/requirements.txt
fastapi
uvicorn
python-multipart
pydub
openai-whisper
torch